+++ vars/vars.go
@@ -3,7 +3,7 @@ package vars
 // Version info
 var (
 	AppName    = "go-oak-chunk"
-	AppVersion = "2.0"
+	AppVersion = "3.1.0"
 	GoVersion  = "default"
 	BuildTime  = "default"
 	GitCommit  = "default"